+++ src/app/api/route.ts
@@ -1,0 +1,6 @@
+export function GET() {
+	return Response.json({
+		healthcheck: true,
+		message: "Hello from the other side...!",
+	});
+}